Move test base classes to TestExtensions
