+++ test/Tester/CollectionFixtures.cs
@@ -2,7 +2,7 @@
 
 namespace Tester
 {
-    // Assembly collections might be defined once in each assembly
+    // Assembly collections must be defined once in each assembly
     [CollectionDefinition("DefaultCluster")]
     public class DefaultClusterTestCollection : ICollectionFixture<DefaultClusterFixture> { }
 }
