+++ test/Tester/Tester.csproj
@@ -49,7 +49,6 @@
     <Compile Include="DeploymentTests\DeployTest.cs" />
     <Compile Include="Forwarding\ShutdownSiloTests.cs" />
     <Compile Include="MethodInterceptionTests.cs" />
-    <Compile Include="BaseClusterFixture.cs" />
     <Compile Include="General\JsonGrainTests.cs" />
     <Compile Include="PSClientTests\PSClientTests.cs" />
     <Compile Include="RelationalUtilities\PostgreSqlStorageForTesting.cs" />
@@ -81,19 +80,16 @@
     <Compile Include="GrainActivateDeactivateTests.cs" />
     <Compile Include="GrainFactoryTests.cs" />
     <Compile Include="GrainInterfaceHierarchyTests.cs" />
-    <Compile Include="HostedTestClusterBase.cs" />
     <Compile Include="KeyExtensionTests.cs" />
     <Compile Include="ManagementGrainTests.cs" />
     <Compile Include="MembershipTests\LivenessTests.cs" />
     <Compile Include="MembershipTests\SilosStopTests.cs" />
     <Compile Include="MemoryStorageProviderTests.cs" />
     <Compile Include="ObserverTests.cs" />
-    <Compile Include="OrleansTestingBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="..\..\src\Build\GlobalAssemblyInfo.cs">
       <Link>Properties\GlobalAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="DefaultClusterFixture.cs" />
     <Compile Include="ProviderTests.cs" />
     <Compile Include="RelationalUtilities\MySqlStorageForTesting.cs" />
     <Compile Include="InvocationInterceptTests.cs" />
