+++ test/TesterInternal/CollectionFixtures.cs
@@ -3,7 +3,7 @@ using Xunit;
 
 namespace UnitTests
 {
-    // Assembly collections might be defined once in each assembly
+    // Assembly collections must be defined once in each assembly
     [CollectionDefinition("DefaultCluster")]
     public class DefaultClusterTestCollection : ICollectionFixture<DefaultClusterFixture> { }
 }
